Text.RE.REOptions

The Options Tutorial

This API module provides the generic types used to specify the options when compiling REs for each of the backl ends.

See the tutorials at http://re-tutorial-options.regex.uk

SimpleREOptions

data SimpleREOptions

the default API uses these simple, universal RE options, which get auto-converted into the appropriate back-end REOptions_

Constructors

MultilineSensitive
case-sensitive with ^ and $ matching the start and end of a line

MultilineInsensitive
case-insensitive with ^ and $ matsh the start and end of a line

BlockSensitive
case-sensitive with ^ and $ matching the start and end of the input text

BlockInsensitive
case-insensitive with ^ and $ matching the start and end of the input text

REOptions_

data REOptions_ r c e

the general options for an RE are dependent on which back end is being used and are parameterised over the RE type for the back end, and its CompOption and ExecOption types (the compile-time and execution time options, respectively); each back end will define an REOptions type that fills out these three type parameters with the appropriate types (see, for example, Text.RE.TDFA)

Constructors

REOptions

Fields

optionsMacs :: !(Macros r)

the available TestBench RE macros

optionsComp :: !c

the back end compile-time options

optionsExec :: !e

the back end execution-time options

The Macro Tables

type Macros r = HashMap MacroID r

our macro tables are parameterised over the back end RE type and and just associate each MacroID with an RE (which may in turn contain macros to be expanded)

newtype MacroID

MacroID is just a wrapped String type with an IsString instance

Constructors

MacroID

Fields

getMacroID :: String

emptyMacros :: Macros r

a macro table containing no entries