Text.RE.Replace

The Replacing Tutorial

This API module covers the specialised regex tools for doing general editing on text, including the internal details of the Matches and Match types and the associated functions for extracting captures and applying functions to them to transform the subject text.

See the tutorials at http://re-tutorial-replacing.regex.uk

replaceAll

replaceAll :: Replace a => a -> Matches a -> a

replace all with a template, $0 for whole text, $1 for first capture, etc.

replaceAllCaptures :: Replace a => REContext -> (Match a -> RELocation -> Capture a -> Maybe a) -> Matches a -> a

substitutes using a function that takes the full Match context and returns the same replacement text as the _phi_phi context.

replaceAllCaptures_ :: Extract a => ReplaceMethods a -> REContext -> (Match a -> RELocation -> Capture a -> Maybe a) -> Matches a -> a

replaceAllCaptures_ is like like replaceAllCaptures but takes the Replace methods through the ReplaceMethods argument

replaceAllCapturesM :: (Extract a, Monad m) => ReplaceMethods a -> REContext -> (Match a -> RELocation -> Capture a -> m (Maybe a)) -> Matches a -> m a

replaceAllCapturesM is just a monadically generalised version of replaceAllCaptures_

replace

replace :: Replace a => a -> Match a -> a

replaceCaptures :: Replace a => REContext -> (Match a -> RELocation -> Capture a -> Maybe a) -> Match a -> a

substitutes using a function that takes the full Match context and returns the same replacement text as the _phi_phi context.

replaceCaptures_ :: Extract a => ReplaceMethods a -> REContext -> (Match a -> RELocation -> Capture a -> Maybe a) -> Match a -> a

replaceCaptures_ is like replaceCaptures but takes the Replace methods through the ReplaceMethods argument

replaceCapturesM :: (Monad m, Extract a) => ReplaceMethods a -> REContext -> (Match a -> RELocation -> Capture a -> m (Maybe a)) -> Match a -> m a

replaceCapturesM is just a monadically generalised version of replaceCaptures_

REContext and RELocation

data REContext

REContext specifies which contexts the substitutions should be applied

Constructors

TOP
substitutions should be applied to the top-level only, the text that matched the whole RE

SUB
substitutions should only be applied to the text captured by bracketed sub-REs

ALL
the substitution function should be applied to all captures, the top level and the sub-expression captures

data RELocation

the RELocation information passed into the substitution function specifies which sub-expression is being substituted

Constructors

RELocation

Fields

locationMatch :: Int

the zero-based, i-th string to be matched, when matching all strings, zero when only the first string is being matched

locationCapture :: CaptureOrdinal

0, when matching the top-level string matched by the whole RE, 1 for the top-most, left-most redex captured by bracketed sub-REs, etc.

isTopLocation :: RELocation -> Bool

True iff the location references a complete match (i.e., not a bracketed capture)

Matches

data Matches a

the result of matching a RE against a text (with *=~), retaining the text that was matched against

Constructors

Matches

Fields

matchesSource :: !a

the source text being matched

allMatches :: ![Match a]

all Match instances found, left to right

anyMatches :: Matches a -> Bool

tests whether the RE matched the source text at all

countMatches :: Matches a -> Int

count the matches

matches :: Matches a -> [a]

list the texts that Matched

mainCaptures :: Matches a -> [Capture a]

extract the main capture from each match

Match

data Match a

the result of matching a RE to a text once (with ?=~), retaining the text that was matched against

Constructors

Match

Fields

matchSource :: !a

the whole source text

captureNames :: !CaptureNames

the RE's capture names

matchArray :: !(Array CaptureOrdinal (Capture a))

.n-1 captures, starting with the text matched by the whole RE

noMatch :: a -> Match a

Construct a Match that does not match anything.

emptyMatchArray :: Array CaptureOrdinal (Capture a)

an empty array of Capture

matched :: Match a -> Bool

tests whether the RE matched the source text at all

matchedText :: Match a -> Maybe a

yields the text matched by the RE, Nothing if no match

matchCapture :: Match a -> Maybe (Capture a)

the top-level capture if the source text matched the RE, Nothing otherwise

matchCaptures :: Match a -> Maybe (Capture a, [Capture a])

the main top-level capture (capture '0'') and the sub captures if the text matched the RE, Nothing otherwise

(!$$) :: Match a -> CaptureID -> a infixl 9

an alternative for captureText

captureText :: CaptureID -> Match a -> a

look up the text of the nth capture, 0 being the match of the whole RE against the source text, 1, the first bracketed sub-expression to be matched and so on

(!$$?) :: Match a -> CaptureID -> Maybe a

an alternative for captureTextMaybe

captureTextMaybe :: CaptureID -> Match a -> Maybe a

look up the text of the nth capture (0 being the match of the whole), returning Nothing if the Match doesn't contain the capture

(!$) :: Match a -> CaptureID -> Capture a infixl 9

an alternative for capture

capture :: CaptureID -> Match a -> Capture a

look up the nth capture, 0 being the match of the whole RE against the source text, 1, the first bracketed sub-expression to be matched and so on

(!$?) :: Match a -> CaptureID -> Maybe (Capture a)

an alternative for capture captureMaybe

captureMaybe :: CaptureID -> Match a -> Maybe (Capture a)

look up the nth capture, 0 being the match of the whole RE against the source text, 1, the first bracketed sub-expression to be matched and so on, returning Nothing if there is no such capture, or if the capture failed to capture anything (being in a failed alternate)

convertMatchText :: RegexFix regex source => regex -> source -> MatchText source -> Match source

convert a regex-base native MatchText into a regex Match type

Capture

data Capture a

the matching of a single sub-expression against part of the source text

Constructors

Capture

Fields

captureSource :: !a

the whole text that was searched

capturedText :: !a

the text that was matched

captureOffset :: !Int

the number of characters preceding the match with -1 used if no text was captured by the RE (not even the empty string)

captureLength :: !Int

the number of chacter in the captured sub-string

hasCaptured :: Capture a -> Bool

test if the capture has matched any text

capturePrefix :: Extract a => Capture a -> a

returns the text preceding the match

captureSuffix :: Extract a => Capture a -> a

returns the text after the match

CaptureID

data CaptureID

CaptureID identifies captures, either by number (e.g., [cp|1|]) or name (e.g., [cp|foo|]).

Constructors

IsCaptureOrdinal CaptureOrdinal

IsCaptureName CaptureName

type CaptureNames = HashMap CaptureName CaptureOrdinal

the dictionary for named captures stored in compiled regular expressions associates

noCaptureNames :: CaptureNames

an empty CaptureNames dictionary

newtype CaptureName

a CaptureName is just the text of the name

Constructors

CaptureName

Fields

getCaptureName :: Text

newtype CaptureOrdinal

a CaptureOrdinal is just the number of the capture, starting with 0 for the whole of the text matched, then in leftmost, outermost

Constructors

CaptureOrdinal

Fields

getCaptureOrdinal :: Int

findCaptureID :: CaptureID -> CaptureNames -> Either String CaptureOrdinal

look up a CaptureID in the CaptureNames dictionary

Replace and ReplaceMethods

class (Show a, Eq a, Ord a, Extract a, Monoid a) => Replace a where

Replace provides the missing needed to replace the matched text in a Replace a => Match a.

Minimal complete definition

lengthR, packR, unpackR, linesR, unlinesR, parseTemplateR

Methods

lengthR :: a -> Int

length function for a

packR :: String -> a

inject String into a

unpackR :: a -> String

project a onto a String

textifyR :: a -> Text

inject into Text

detextifyR :: Text -> a

project Text onto a

linesR :: a -> [a]

split into lines

unlinesR :: [a] -> a

concatenate a list of lines

appendNewlineR :: a -> a

append a newline

substR :: (a -> a) -> Capture a -> a

apply a substitution function to a Capture

parseTemplateR :: a -> Match a -> RELocation -> Capture a -> Maybe a

convert a template containing $0, $1, etc., in the first argument, into a phi replacement function for use with replaceAllCaptures and replaceCaptures

data ReplaceMethods a

a selection of the Replace methods can be encapsulated with ReplaceMethods for the higher-order replacement functions

Constructors

ReplaceMethods

Fields

methodLength :: a -> Int

methodSubst :: (a -> a) -> Capture a -> a

replaceMethods :: Replace a => ReplaceMethods a

replaceMethods encapsulates ReplaceMethods a from a Replace a context