Text.RE.TestBench

The Test Bench Tutorial

This API module provides a test bench for developing, documenting and testing regex RE macros.

See the tutorials at http://re-tutorial-testbench.regex.uk

The Test Bench

type MacroEnv = HashMap MacroID MacroDescriptor

each macro can reference others, the whole environment being required for each macro, so we use a Lazy HashMap

data MacroDescriptor

describes a macro, giving the text of the RE and a summary description

Constructors

MacroDescriptor

Fields

macroSource :: !RegexSource

the RE

macroSamples :: ![String]

some sample matches

macroCounterSamples :: ![String]

some sample non-matches

macroTestResults :: ![TestResult]

validation test results

macroParser :: !(Maybe FunctionID)

WA, the parser function

macroDescription :: !String

summary comment

newtype RegexSource

a RE that should work for POSIX and PCRE with open brackets ('(') represented as follows: ( mere symbol (?: used for grouping only, not for captures (}: used for captures only, not for grouping (]: used for captures and grouping ( do not modify

Constructors

RegexSource

Fields

_RegexSource :: String

data WithCaptures

do we need the captures in the RE or would they be stripped out where possible

Constructors

InclCaptures
include all captures

ExclCaptures
remove captures where possible

data RegexType

what flavour of regex are we dealing with

isTDFA :: RegexType -> Bool

test RegexType for TDFA/PCREness

isPCRE :: RegexType -> Bool

test RegexType for TDFA/PCREness

presentRegexType :: RegexType -> String

Constructing a MacrosEnv

mkMacros :: (Monad m, Functor m) => (String -> m r) -> RegexType -> WithCaptures -> MacroEnv -> m (Macros r)

construct a macro table suitable for use with the RE compilers

Formatting Macros

formatMacroTable :: RegexType -> MacroEnv -> String

format a macros table as a markdown table

formatMacroSummary :: RegexType -> MacroEnv -> MacroID -> String

generate a plain text summary of a macro

formatMacroSources :: RegexType -> WithCaptures -> MacroEnv -> String

list the source REs for each macro in plain text

formatMacroSource :: RegexType -> WithCaptures -> MacroEnv -> MacroID -> String

list the source of a single macro in plain text

mdRegexSource :: RegexType -> WithCaptures -> MacroEnv -> MacroDescriptor -> String

Formatting Macros

testMacroEnv :: String -> RegexType -> MacroEnv -> IO Bool

test that a MacroEnv is passing all of its built-in tests

runTests :: (Eq a, Show a) => RegexType -> (String -> Maybe a) -> [(String, a)] -> MacroEnv -> MacroID -> MacroDescriptor -> MacroDescriptor

runTests' :: (Eq a, Show a) => RegexType -> (Match String -> Maybe a) -> [(String, a)] -> MacroEnv -> MacroID -> MacroDescriptor -> MacroDescriptor

The Parsers

module Text.RE.TestBench.Parsers

The Match Type

data Match a

the result of matching a RE to a text once (with ?=~), retaining the text that was matched against