Text.RE.Tools.Grep

Documentation

The Grep toolkit matches REs against each line of a text.

See the Regex Tools tutorial at http://re-tutorial-tools.regex.uk

grep :: IsRegex re ByteString => Verbosity -> re -> FilePath -> IO ()

operates a bit like classic grep printing out the lines matched

data Verbosity

specifies whether to return the lines matched or missed

Constructors

LinesMatched

LinesNotMatched

data Line s

grepLines returns a Line for each line in the file, listing all of the Matches for that line

Constructors

Line

Fields

getLineNumber :: LineNo

the LineNo for this line

getLineMatches :: Matches s

all the Matches of the RE on this line

grepLines :: IsRegex re ByteString => re -> FilePath -> IO [Line ByteString]

returns a Line for each line in the file, enumerating all of the matches for that line

grepFilter :: IsRegex re s => re -> s -> [Line s]

returns a Line for each line in the argument text, enumerating all of the matches for that line

type GrepScript re s t = [(re, LineNo -> Matches s -> Maybe t)]

a GrepScript lists RE-action associations, with the first RE to match a line selecting the action to be executed on each line in the file

grepWithScript :: IsRegex re s => GrepScript re s t -> [s] -> [t]

given a list of lines, apply the GrepScript to each line of the file

report :: Verbosity -> [Line ByteString] -> String

generate a grep report from a list of Line

linesMatched :: Verbosity -> [Line s] -> [Line s]

given a velocity flag filter out either the lines matched or not matched

IsRegex

class Replace s => IsRegex re s where

the IsRegex class allows polymorhic tools to be written that will work with a variety of regex back ends and text types

Minimal complete definition

matchOnce, matchMany, makeRegexWith, makeSearchReplaceWith, regexSource

Methods

matchOnce :: re -> s -> Match s

finding the first match

matchMany :: re -> s -> Matches s

finding all matches

makeRegex :: (Functor m, Monad m, MonadFail m) => s -> m re

compiling an RE, failing if the RE is not well formed

makeRegexWith :: (Functor m, Monad m, MonadFail m) => SimpleREOptions -> s -> m re

comiling an RE, specifying the SimpleREOptions

makeSearchReplace :: (Functor m, Monad m, MonadFail m, IsRegex re s) => s -> s -> m (SearchReplace re s)

compiling a SearchReplace template from the RE text and the template Text, failing if they are not well formed

makeSearchReplaceWith :: (Functor m, Monad m, MonadFail m, IsRegex re s) => SimpleREOptions -> s -> s -> m (SearchReplace re s)

compiling a SearchReplace template specifying the SimpleREOptions for the RE

makeEscaped :: (Functor m, Monad m, MonadFail m) => (s -> s) -> s -> m re

incorporate an escaped string into a compiled RE with the default options

makeEscapedWith :: (Functor m, Monad m, MonadFail m) => SimpleREOptions -> (s -> s) -> s -> m re

incorporate an escaped string into a compiled RE with the specified SimpleREOptions

regexSource :: re -> s

extract the text of the RE from the RE

data SearchReplace re s

contains a compiled RE and replacement template

Constructors

SearchReplace

Fields

getSearch :: !re

the RE to match a string to replace

getTemplate :: !s

the replacement template with ${cap} used to identify a capture (by number or name if one was given) and $$ being used to escape a single $

searchReplaceAll :: IsRegex re s => SearchReplace re s -> s -> s

search and replace all matches in the argument text; e.g., this function will convert every YYYY-MM-DD format date in its argument text into a DD/MM/YYYY date:

searchReplaceAll [ed|${y}([0-9]{4})-0*${m}([0-9]{2})-0*${d}([0-9]{2})///${d}/${m}/${y}|]

searchReplaceFirst :: IsRegex re s => SearchReplace re s -> s -> s

search and replace the first occurrence only (if any) in the input text e.g., to prefix the first string of four hex digits in the imput text, if any, with 0x:

searchReplaceFirst [ed|[0-9A-Fa-f]{4}///0x$0|]

LineNo

newtype LineNo

our line numbers are of the proper zero-based kind

Constructors

ZeroBasedLineNo

Fields

getZeroBasedLineNo :: Int

firstLine :: LineNo

the first line in a file

getLineNo :: LineNo -> Int

extract a conventional 1-based line number

lineNo :: Int -> LineNo

inject a conventional 1-based line number

Replace

module Text.RE.Replace