Text.RE.Tools.Lex

Find

The Lex toolkit uses REs to identify tokens in a file, returning a list of tokens.

See the Regex Tools tutorial at http://re-tutorial-tools.regex.uk

alex :: IsRegex re s => [(re, Match s -> Maybe t)] -> t -> s -> [t]

a simple regex-based scanner interpreter for prototyping scanners

alex' :: Replace s => (re -> s -> Match s) -> [(re, Match s -> Maybe t)] -> t -> s -> [t]

a higher order version of alex parameterised over the matchOnce function

IsRegex

class Replace s => IsRegex re s where

the IsRegex class allows polymorhic tools to be written that will work with a variety of regex back ends and text types

Minimal complete definition

matchOnce, matchMany, makeRegexWith, makeSearchReplaceWith, regexSource

Methods

matchOnce :: re -> s -> Match s

finding the first match

matchMany :: re -> s -> Matches s

finding all matches

makeRegex :: (Functor m, Monad m, MonadFail m) => s -> m re

compiling an RE, failing if the RE is not well formed

makeRegexWith :: (Functor m, Monad m, MonadFail m) => SimpleREOptions -> s -> m re

comiling an RE, specifying the SimpleREOptions

makeSearchReplace :: (Functor m, Monad m, MonadFail m, IsRegex re s) => s -> s -> m (SearchReplace re s)

compiling a SearchReplace template from the RE text and the template Text, failing if they are not well formed

makeSearchReplaceWith :: (Functor m, Monad m, MonadFail m, IsRegex re s) => SimpleREOptions -> s -> s -> m (SearchReplace re s)

compiling a SearchReplace template specifying the SimpleREOptions for the RE

makeEscaped :: (Functor m, Monad m, MonadFail m) => (s -> s) -> s -> m re

incorporate an escaped string into a compiled RE with the default options

makeEscapedWith :: (Functor m, Monad m, MonadFail m) => SimpleREOptions -> (s -> s) -> s -> m re

incorporate an escaped string into a compiled RE with the specified SimpleREOptions

regexSource :: re -> s

extract the text of the RE from the RE

data SearchReplace re s

contains a compiled RE and replacement template

Constructors

SearchReplace

Fields

getSearch :: !re

the RE to match a string to replace

getTemplate :: !s

the replacement template with ${cap} used to identify a capture (by number or name if one was given) and $$ being used to escape a single $

searchReplaceAll :: IsRegex re s => SearchReplace re s -> s -> s

search and replace all matches in the argument text; e.g., this function will convert every YYYY-MM-DD format date in its argument text into a DD/MM/YYYY date:

searchReplaceAll [ed|${y}([0-9]{4})-0*${m}([0-9]{2})-0*${d}([0-9]{2})///${d}/${m}/${y}|]

searchReplaceFirst :: IsRegex re s => SearchReplace re s -> s -> s

search and replace the first occurrence only (if any) in the input text e.g., to prefix the first string of four hex digits in the imput text, if any, with 0x:

searchReplaceFirst [ed|[0-9A-Fa-f]{4}///0x$0|]

Replace

module Text.RE.Replace