Text.RE.ZeInternals

The regex Internal Modules

This module contains just what the test suite (re-tests) in regex-examples needs from the package internals to do its job and the ZeInternals types and functions needed by the regex-with-pcre package

Text.RE.ZeInternals.AddCaptureNames

addCaptureNames :: Typeable a => CaptureNames -> a -> a

a hairy dynamically-typed function used with the legacy (=~) and (=~~) to see if it can/should add the capture names extracted from the RE into the polymorphic result of the operator (it does for any Match or Matches type, provided it is parameterised over a recognised type). The test suite is all over this one, testing all of these cases.

addCaptureNamesToMatches :: CaptureNames -> Matches a -> Matches a

a convenience function used by the API modules to insert capture names extracted from the parsed RE into the (*=~) result

addCaptureNamesToMatch :: CaptureNames -> Match a -> Match a

a convenience function used by the API modules to insert capture names extracted from the parsed RE into the (?=~) result

Text.RE.ZeInternals.EscapeREString

escapeREString :: String -> String

Convert a string into a regular expression that will match that string

Text.RE.ZeInternals.NamedCaptures

cp :: QuasiQuoter

quasi quoter for CaptureID: [cp|0|], [cp|0|], etc., indexing captures by classic positional numbers, and [cp|foo|], etc., referencing a named capture [re| ... ${foo}( ... ) ... |].

extractNamedCaptures :: String -> Either String ((Int, CaptureNames), String)

extract the CaptureNames from an RE or return an error diagnostic if the RE is not well formed; also returns the total number of captures in the RE

idFormatTokenREOptions :: FormatTokenREOptions

a configuration that will preserve the parsed regular expression in the output

data Token

our RE scanner returns a list of these tokens

validToken :: Token -> Bool

check that a token is well formed

formatTokens :: [Token] -> String

format [Token] into an RE string

formatTokens' :: FormatTokenREOptions -> [Token] -> String

the general Token formatter, generating REs according to the options

formatTokens0 :: [Token] -> String

scan :: String -> [Token]

scan a RE string into a list of RE Token

Text.RE.ZeInternals.Replace

expandMacros :: (r -> String) -> Macros r -> String -> String

expand all of the @{..} macros in the RE in the argument String according to the Macros argument, preprocessing the RE String according to the Mode argument (used internally)

Text.RE.ZeInternals.PreludeMacros

data PreludeMacro

an enumeration of all of the prelude macros

Constructors

PM_nat

PM_hex

PM_int

PM_frac

PM_string

PM_string_simple

PM_id

PM_id'

PM_id_

PM_date

PM_date_slashes

PM_time

PM_timezone

PM_datetime

PM_datetime_8601

PM_datetime_clf

PM_shortmonth

PM_address_ipv4

PM_email_simple

PM_url

PM_syslog_severity

presentPreludeMacro :: PreludeMacro -> String

naming the macros

preludeMacros :: (Monad m, Functor m) => (String -> m r) -> RegexType -> WithCaptures -> m (Macros r)

generate the standard prelude Macros used to parse REs

preludeMacroTable :: RegexType -> String

format the standard prelude macros in a markdown table

preludeMacroSummary :: RegexType -> PreludeMacro -> String

generate a textual summary of the prelude macros

preludeMacroSources :: RegexType -> String

generate a plain text table giving the RE for each macro with all macros expanded (to NF)

preludeMacroSource :: RegexType -> PreludeMacro -> String

generate plain text giving the expanded RE for a single macro

preludeMacroEnv :: RegexType -> MacroEnv

generate the MacroEnv for the standard prelude macros

Text.RE.ZeInternals.SearchReplace

unsafeCompileSearchReplace_ :: (String -> s) -> (String -> Either String re) -> String -> SearchReplace re s

warapper on compileSearchReplace_ that will generate an error if any compilation errors are found

compileSearchReplace_ :: (Monad m, MonadFail m, Functor m) => (String -> s) -> (String -> Either String re) -> String -> m (SearchReplace re s)

compile a SearchReplace template generating errors if the RE or the template are not well formed -- all capture references being checked

compileSearchAndReplace_ :: (Monad m, MonadFail m, Functor m) => (String -> s) -> (String -> Either String re) -> String -> String -> m (SearchReplace re s)

compile SearcgReplace from two strings containing the RE and the replacement template

Text.RE.ZeInternals.QQ

data QQFailure

used to throw an exception reporting an abuse of a quasi quoter

Constructors

QQFailure

Fields

_qqf_context :: String

in what context was the quasi quoter used

_qqf_component :: String

how was the quasi quoter being abused

qq0 :: String -> QuasiQuoter

a quasi quoter that can be used in no context (to be extended with the appropriate quasi quoter parser)

Text.RE.ZeInternals.TestBench

mkTDFA :: TestBenchMatcher -> RegexType

mkPCRE :: TestBenchMatcher -> RegexType

badMacros :: MacroEnv -> [MacroID]